chore(web): reorganizes gesture-recognizer source in prep for chopping subsegmentation
