+++ common/web/gesture-recognizer/package.json
@@ -27,12 +27,8 @@
     "typescript": "^4.9.5"
   },
   "scripts": {
-    "build": "echo 'Building @keymanapp/gesture-recognizer' && tsc -b ./tsconfig.json",
-    "clean": "tsc -b --clean",
-    "karma": "karma",
-    "mocha": "mocha",
-    "test": "gosh ./src/test/test.sh",
-    "tsc": "tsc"
+    "build": "gosh ./build.sh",
+    "test": "gosh ./test.sh"
   },
   "dependencies": {
     "eventemitter3": "^4.0.0"
