+++ common/web/gesture-recognizer/src/engine/configuration/gestureRecognizerConfiguration.ts
