+++ common/web/gesture-recognizer/src/engine/configuration/recognitionZoneSource.ts
