+++ common/web/gesture-recognizer/src/engine/configuration/viewportZoneSource.ts
