+++ common/web/gesture-recognizer/src/engine/headless/inputSample.ts
