+++ common/web/gesture-recognizer/src/engine/headless/segment.ts
