+++ common/web/gesture-recognizer/src/engine/headless/trackedPath.ts
@@ -1,6 +1,6 @@
 import EventEmitter from "eventemitter3";
 import { InputSample } from "./inputSample.js";
-import { PathSegmenter } from "./pathSegmenter.js";
+import { PathSegmenter } from "./subsegmentation/pathSegmenter.js";
 import { Segment } from "./segment.js";
 
 /**
