+++ common/web/gesture-recognizer/src/engine/index.ts
@@ -1,0 +1,18 @@
+export { ConstructingSegment } from './headless/subsegmentation/constructingSegment.js';
+export { CumulativePathStats } from './headless/cumulativePathStats.js';
+export { GestureRecognizer } from "./gestureRecognizer.js";
+export { GestureRecognizerConfiguration } from "./configuration/gestureRecognizerConfiguration.js"
+export { InputSample } from "./headless/inputSample.js";
+export { JSONTrackedInput, TrackedInput } from "./trackedInput.js";
+export { JSONTrackedPath, TrackedPath } from "./headless/trackedPath.js";
+export { JSONTrackedPoint, TrackedPoint } from "./trackedPoint.js";
+export { MouseEventEngine } from "./mouseEventEngine.js";
+export { PathSegmenter, Subsegmentation } from "./headless/subsegmentation/pathSegmenter.js";
+export { PaddedZoneSource } from './configuration/paddedZoneSource.js';
+export { RecognitionZoneSource } from './configuration/recognitionZoneSource.js';
+export { Segment } from "./headless/segment.js";
+export { SegmentClassifier } from "./headless/segmentClassifier.js";
+export { TouchEventEngine } from "./touchEventEngine.js";
+export { ViewportZoneSource } from './configuration/viewportZoneSource.js';
+
+// TODO:  export other odds & ends.
