--- common/web/gesture-recognizer/src/index.ts
@@ -1,15 +1,0 @@
-export { ConstructingSegment } from './constructingSegment.js';
-export { CumulativePathStats } from './cumulativePathStats.js';
-export { GestureRecognizer } from "./gestureRecognizer.js";
-export { GestureRecognizerConfiguration } from "./gestureRecognizerConfiguration.js"
-export { InputSample } from "./inputSample.js";
-export { JSONTrackedInput, TrackedInput } from "./trackedInput.js";
-export { JSONTrackedPath, TrackedPath } from "./trackedPath.js";
-export { JSONTrackedPoint, TrackedPoint } from "./trackedPoint.js";
-export { MouseEventEngine } from "./mouseEventEngine.js";
-export { PathSegmenter, Subsegmentation } from "./pathSegmenter.js";
-export { Segment } from "./segment.js";
-export { SegmentClassifier } from "./segmentClassifier.js";
-export { TouchEventEngine } from "./touchEventEngine.js";
-
-// TODO:  export other odds & ends.
