+++ common/web/gesture-recognizer/src/test/auto/headless/trackedPath.js
@@ -1,13 +1,17 @@
 import { assert } from 'chai'
 import sinon  from 'sinon';
 import fs from 'fs';
+import path from 'path';
+import url from 'url';
 
 import { TrackedPath } from '@keymanapp/gesture-recognizer';
 
 import { HeadlessRecordingSimulator, timedPromise } from '../../../../build/tools/obj/index.js';
 
-// For the integrated-style recording-based test.
-const SEGMENT_TEST_JSON_FOLDER = './resources/json/segmentation';
+// Ensures that the resources are resolved relative to this script, not to the cwd when the test
+// runner was launched.
+const scriptFolder = path.dirname(url.fileURLToPath(import.meta.url));
+const SEGMENT_TEST_JSON_FOLDER = path.resolve(`${scriptFolder}/../../resources/json/segmentation`);
 
 import { assertSegmentSimilarity } from '../../resources/assertSegmentSimilarity.js'
 ;
