+++ common/web/gesture-recognizer/test.sh
@@ -3,7 +3,7 @@
 ## START STANDARD BUILD SCRIPT INCLUDE
 # adjust relative paths as necessary
 THIS_SCRIPT="$(greadlink -f "${BASH_SOURCE[0]}" 2>/dev/null || readlink -f "${BASH_SOURCE[0]}")"
-. "$(dirname "$THIS_SCRIPT")/../../../../../resources/build/build-utils.sh"
+. "$(dirname "$THIS_SCRIPT")/../../../resources/build/build-utils.sh"
 ## END STANDARD BUILD SCRIPT INCLUDE
 
 . $KEYMAN_ROOT/resources/shellHelperFunctions.sh
@@ -44,7 +44,7 @@ test-headless ( ) {
     MOCHA_FLAGS="--reporter mocha-teamcity-reporter"
   fi
 
-  mocha --recursive $MOCHA_FLAGS ./auto/headless/
+  c8 mocha --recursive $MOCHA_FLAGS ./src/test/auto/headless/
 }
 
 test-browser ( ) {
@@ -58,7 +58,7 @@ test-browser ( ) {
     KARMA_CONFIG="CI.conf.cjs"
   fi
 
-  karma start auto/browser/$KARMA_CONFIG "$KARMA_FLAGS"
+  karma start src/test/auto/browser/$KARMA_CONFIG "$KARMA_FLAGS"
 }
 
 if builder_start_action test:headless; then
